requirements.txt
keras
numpy
Pillow
opencv-python-headless
numpy
tensorflow


